finish support for create
The arch system section should be separate from the os.
This change will fully read in a compatibiity request (yaml),
and then map requested fields to extractor output fields,
and with the option to add on-the-fly custom metadata.
We output a finished compatibility artifact that is ready
to be pushed to a registry and later used paired with an
image. Note that I need to run a final jsonSchema validate
on generation (before save) still.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>