add example prototype for check
This update will allow reading in a manifest, which is a list of
pairs of images and artifacts. Thanks to the media type of the
artifact layer we can easily find the compatibility spec and then
load it into the library. This is the core functionaltiy that
we need to next create a graph to query over, and likely
taking user preferences (a host to match to). While I could
implement this very simply, I want to go down a bit of a rabbit
hole and experiment with making a graph library/interface in
go akin to kraft. Yeah, I kind of just want to have fun. ;)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>