add start of parsing schemas and artifacts into graph
This will generate the first schemas graph, and then prototype adding a single
artifact (image reference and compatibility schema) onto it. Some design choices
that I made are to store the mapping separate from the graph, that way we can
always nuke them and start fresh without needing to recreate the graph. I also
model a value (the leaf of the tree) as just another node, the idea being that
we could extend the graph to include some new children. This sort of breaks
the model that only the schemas should be nodes, but it is not so terrible an
idea because in theory adding a value as a node on the fly could represent
an updated cluster state (that is not present in the schema yet). THis
might not be perfect but is a great start! I will next try to extract more
of the artifacts and run this for a more reasonable image selection test
case.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>