finishing up match command!
This will complete the algorithm, which is very simple for now.
We build the graph from schemas, and then with depth first search
map metadata features to it. Instead of needing to search again,
we simply keep a hash of image identifiers at each node (e.g.,
where a node is a metadata attribute) and then we can take
a request, look up each feature node in N(1) time, and then
take the intersection of images across those nodes to get
the final matching set! This will work well for small
graphs, as long as memory is not an issue to store that
hash. For larger graphs we may need to do an actual
search.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>