feat: add creator plugins
Problem: we did not have a way to define a creator as a plugin.
Solution: add a plugin interface to create. I originally was going
to create separate plugin types, but I like the idea that one plugin
family can decide to define both easily. When this is refactored to
have a more "register" design (to make it flexible to changing the set
available) it will be nice to provide Create/Extract from the same
interface and keep the number of interfaces / functions for them
minimal. I was going to add hwloc now, but there seems to be a
bug so we will need to move forward prototyping with the current
proxy for nodes, which is just using the go runtime package. We
obviously need to improve upon this.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>