wip: integration with rainbow
This set of small changes is to support node generation for the
rainbow scheduler. There was a small bug in the way I was serializing
the cluster graph that it forgot the top level "graph" key, and
I have also cleaned up passing of arguments for the plugins. The paths
for containment also were slightly off. I will leave this PR open as
I figure out how to actually accept the graph in rainbow. I wanted
to use fluxion but I realize there is not support for extending
or growing a graph so I would not be able to add more than one cluster.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>